fix: complete durable human approval review flow
