+++ contracts/functional-scenario-evidence.json
@@ -45,6 +45,13 @@
       ]
     },
     {
+	  "test": "tests/functional/transport/cli/commands/human_approval_wiring_test.go::TestCLIWorkApprovalListAndShowExposePendingApprovalAndSafeEmptyErrors",
+	  "stableIds": [
+	    "cli/you.work.approval.list",
+	    "cli/you.work.approval.show"
+	  ]
+	},
+	{
       "test": "tests/functional/transport/cli/commands/run_wiring_test.go::TestCLIRunFactoryByPath",
       "stableIds": [
         "cli/you.run"
