--- pkg/services/factory_definitions/definition/external_test_helpers_test.go
@@ -1,5 +1,0 @@
-package factorydefinition_test
-
-func definitionStringPtr(value string) *string {
-	return &value
-}
