--- pkg/services/factory_definitions/definition/internal_test_helpers_test.go
@@ -1,5 +1,0 @@
-package factorydefinition
-
-func definitionStringPtr(value string) *string {
-	return &value
-}
