+++ pkg/services/factory_definitions/definition/persistence_test_helper_test.go
@@ -96,3 +96,7 @@ func preparedLayoutPersistence() factorydefinitions.Persistence {
 	}
 	return persistence
 }
+
+func definitionStringPtr(value string) *string {
+	return &value
+}
