+++ pkg/services/factory_definitions/internal/contracts/factory_runtime.go
@@ -1,12 +1,115 @@
 package factorycontracts
 
 import (
+	"fmt"
+	"math"
+	"strings"
 	"time"
 
 	workerconfig "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/validation/authoredmodel/workers"
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 )
+
+const (
+	FactoryWebhookEventTypeWorkStateChange     = "WORK_STATE_CHANGE"
+	FactoryWebhookEventTypeDispatchResponse    = "DISPATCH_RESPONSE"
+	FactoryWebhookEventTypeDispatchReconciled  = "DISPATCH_RECONCILED"
+	FactoryWebhookEventTypeDispatchInterrupted = "DISPATCH_INTERRUPTED"
+
+	FactoryWebhookDispatchStatusFailed      = "FAILED"
+	FactoryWebhookDispatchStatusInterrupted = "INTERRUPTED"
+
+	DefaultFactoryWebhookRequestTimeout    = 10 * time.Second
+	DefaultFactoryWebhookMaxAttempts       = 5
+	DefaultFactoryWebhookInitialBackoff    = time.Second
+	DefaultFactoryWebhookBackoffMultiplier = 2.0
+	DefaultFactoryWebhookMaxBackoff        = 30 * time.Second
+)
+
+// FactoryWebhookConfig declares one outbound subscription without carrying
+// resolved secret material. Delivery policy is resolved only at runtime.
+type FactoryWebhookConfig struct {
+	Name             string                              `json:"name" yaml:"name"`
+	Enabled          bool                                `json:"enabled" yaml:"enabled"`
+	URL              string                              `json:"url" yaml:"url"`
+	SigningSecretRef string                              `json:"signingSecretRef" yaml:"signingSecretRef"`
+	Filter           FactoryWebhookFilterConfig          `json:"filter" yaml:"filter"`
+	DeliveryPolicy   *FactoryWebhookDeliveryPolicyConfig `json:"deliveryPolicy,omitempty" yaml:"deliveryPolicy,omitempty"`
+}
+
+// FactoryWebhookFilterConfig selects canonical Factory Event types and, for
+// dispatch event types, optional canonical dispatch statuses.
+type FactoryWebhookFilterConfig struct {
+	EventTypes       []string `json:"eventTypes" yaml:"eventTypes"`
+	DispatchStatuses []string `json:"dispatchStatuses,omitempty" yaml:"dispatchStatuses,omitempty"`
+}
+
+// FactoryWebhookDeliveryPolicyConfig keeps optional authored values distinct
+// from their effective defaults so explicit invalid zero values are rejected.
+type FactoryWebhookDeliveryPolicyConfig struct {
+	RequestTimeout    *string  `json:"requestTimeout,omitempty" yaml:"requestTimeout,omitempty"`
+	MaxAttempts       *int     `json:"maxAttempts,omitempty" yaml:"maxAttempts,omitempty"`
+	InitialBackoff    *string  `json:"initialBackoff,omitempty" yaml:"initialBackoff,omitempty"`
+	BackoffMultiplier *float64 `json:"backoffMultiplier,omitempty" yaml:"backoffMultiplier,omitempty"`
+	MaxBackoff        *string  `json:"maxBackoff,omitempty" yaml:"maxBackoff,omitempty"`
+}
+
+// FactoryWebhookEffectiveDeliveryPolicy contains parsed, bounded values used
+// by the delivery runtime.
+type FactoryWebhookEffectiveDeliveryPolicy struct {
+	RequestTimeout    time.Duration
+	MaxAttempts       int
+	InitialBackoff    time.Duration
+	BackoffMultiplier float64
+	MaxBackoff        time.Duration
+}
+
+// ResolveFactoryWebhookDeliveryPolicy applies the documented defaults and
+// parses authored Go duration values for a webhook delivery policy.
+func ResolveFactoryWebhookDeliveryPolicy(config *FactoryWebhookDeliveryPolicyConfig) (FactoryWebhookEffectiveDeliveryPolicy, error) {
+	effective := FactoryWebhookEffectiveDeliveryPolicy{RequestTimeout: DefaultFactoryWebhookRequestTimeout, MaxAttempts: DefaultFactoryWebhookMaxAttempts, InitialBackoff: DefaultFactoryWebhookInitialBackoff, BackoffMultiplier: DefaultFactoryWebhookBackoffMultiplier, MaxBackoff: DefaultFactoryWebhookMaxBackoff}
+	if config == nil {
+		return effective, nil
+	}
+	var err error
+	if effective.RequestTimeout, err = resolveFactoryWebhookDuration("requestTimeout", config.RequestTimeout, effective.RequestTimeout); err != nil {
+		return FactoryWebhookEffectiveDeliveryPolicy{}, err
+	}
+	if config.MaxAttempts != nil {
+		if *config.MaxAttempts <= 0 {
+			return FactoryWebhookEffectiveDeliveryPolicy{}, fmt.Errorf("maxAttempts must be positive")
+		}
+		effective.MaxAttempts = *config.MaxAttempts
+	}
+	if effective.InitialBackoff, err = resolveFactoryWebhookDuration("initialBackoff", config.InitialBackoff, effective.InitialBackoff); err != nil {
+		return FactoryWebhookEffectiveDeliveryPolicy{}, err
+	}
+	if config.BackoffMultiplier != nil {
+		if math.IsNaN(*config.BackoffMultiplier) || math.IsInf(*config.BackoffMultiplier, 0) || *config.BackoffMultiplier < 1 {
+			return FactoryWebhookEffectiveDeliveryPolicy{}, fmt.Errorf("backoffMultiplier must be at least 1")
+		}
+		effective.BackoffMultiplier = *config.BackoffMultiplier
+	}
+	if effective.MaxBackoff, err = resolveFactoryWebhookDuration("maxBackoff", config.MaxBackoff, effective.MaxBackoff); err != nil {
+		return FactoryWebhookEffectiveDeliveryPolicy{}, err
+	}
+	if effective.MaxBackoff < effective.InitialBackoff {
+		return FactoryWebhookEffectiveDeliveryPolicy{}, fmt.Errorf("maxBackoff must not be less than initialBackoff")
+	}
+	return effective, nil
+}
+
+func resolveFactoryWebhookDuration(field string, value *string, fallback time.Duration) (time.Duration, error) {
+	if value == nil {
+		return fallback, nil
+	}
+	duration, err := time.ParseDuration(strings.TrimSpace(*value))
+	if err != nil || duration <= 0 {
+		return 0, fmt.Errorf("%s must be a positive Go duration", field)
+	}
+	return duration, nil
+}
 
 // RequestValidationError reports a stable client-side validation failure.
 type RequestValidationError struct {
