+++ pkg/services/factory_sessions/transports/http/handler.go
@@ -8,12 +8,17 @@ package http
 
 import (
 	"context"
+	"errors"
+	"net/http"
+	"strings"
 
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
+	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
+	"github.com/portpowered/infinite-you/pkg/transports/mapping/factorysession"
 	"go.uber.org/zap"
 )
 
@@ -110,3 +115,70 @@ func NewHandler(deps Dependencies, logger *zap.Logger) *Adapter {
 // are kept mechanically identical to the established public behavior.
 type Handler = Adapter
 type Server = Adapter
+
+// ListHumanApprovalsBySessionId returns the pending approvals projected from
+// the selected live Factory Session's canonical event history.
+func (s *Adapter) ListHumanApprovalsBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID, params factoryapi.ListHumanApprovalsBySessionIdParams) {
+	if s.guardSessionsRequestContext(w, r) {
+		return
+	}
+	if params.Status != nil && string(*params.Status) != "PENDING" {
+		s.writeError(w, http.StatusBadRequest, "unsupported human approval status; only PENDING is available", "BAD_REQUEST")
+		return
+	}
+	approvals, err := s.pendingHumanApprovals(r, string(sessionID))
+	if err != nil {
+		if s.writeSessionsRootError(w, string(sessionID), err) {
+			return
+		}
+		s.logger.Error("list human approvals failed", zap.Error(err))
+		s.writeError(w, http.StatusInternalServerError, "failed to list human approvals", "INTERNAL_ERROR")
+		return
+	}
+	s.writeJSON(w, http.StatusOK, factoryapi.ListHumanApprovalsResponse{Approvals: factorysession.HumanApprovalsToAPI(approvals)})
+}
+
+// GetHumanApprovalBySessionId returns one pending approval by its stable
+// identity. Resolution is read-only; decision handling belongs to a later lane.
+func (s *Adapter) GetHumanApprovalBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID, approvalID factoryapi.HumanApprovalID) {
+	if s.guardSessionsRequestContext(w, r) {
+		return
+	}
+	approvals, err := s.pendingHumanApprovals(r, string(sessionID))
+	if err != nil {
+		if s.writeSessionsRootError(w, string(sessionID), err) {
+			return
+		}
+		s.logger.Error("get human approval failed", zap.Error(err))
+		s.writeError(w, http.StatusInternalServerError, "failed to get human approval", "INTERNAL_ERROR")
+		return
+	}
+	for _, approval := range approvals {
+		if approval.ApprovalID == string(approvalID) {
+			s.writeJSON(w, http.StatusOK, factorysession.HumanApprovalToAPI(approval))
+			return
+		}
+	}
+	s.writeError(w, http.StatusNotFound, "human approval not found", "NOT_FOUND")
+}
+
+func (s *Adapter) pendingHumanApprovals(r *http.Request, sessionID string) ([]factorydefinitions.FactoryWorldHumanApproval, error) {
+	if strings.TrimSpace(sessionID) == "" {
+		return nil, errors.New("factory session id is required")
+	}
+	if s.liveControl != nil {
+		projection, err := s.liveControl.GetFactorySession(r.Context(), sessionID)
+		if err != nil {
+			return nil, err
+		}
+		return append([]factorydefinitions.FactoryWorldHumanApproval(nil), projection.Runtime.PendingHumanApprovals...), nil
+	}
+	if s.sessionsRoot != nil {
+		projection, err := s.sessionsRoot.GetFactorySession(r.Context(), sessionID)
+		if err != nil {
+			return nil, err
+		}
+		return append([]factorydefinitions.FactoryWorldHumanApproval(nil), projection.Runtime.PendingHumanApprovals...), nil
+	}
+	return nil, errors.New("factory session read service is unavailable")
+}
