--- pkg/services/recordings/internal/replay/external_test_helpers_test.go
@@ -1,5 +1,0 @@
-package replay_test
-
-func replayStringPtr(value string) *string {
-	return &value
-}
