+++ pkg/transports/cli/commandregistry/work_resolved_list_test.go
@@ -964,172 +964,6 @@ func resolvedTestInputsWithout(
 	return resolvedTestInputs(t, filtered...)
 }
 
-func TestResolvedWorkAdaptersRequireServices(t *testing.T) {
-	tests := []struct {
-		name    string
-		handler commandregistry.ResolvedWorkRunE
-	}{
-		{"approval list", commandregistry.ResolvedApprovalListRunE(commandregistry.ResolvedApprovalListBinding{})},
-		{"approval show", commandregistry.ResolvedApprovalShowRunE(commandregistry.ResolvedApprovalShowBinding{})},
-		{"list", commandregistry.ResolvedListRunE(commandregistry.ResolvedListBinding{})},
-		{"show", commandregistry.ResolvedShowRunE(commandregistry.ResolvedShowBinding{})},
-		{"move", commandregistry.ResolvedMoveRunE(commandregistry.ResolvedMoveBinding{})},
-		{"visualize", commandregistry.ResolvedVisualizeRunE(commandregistry.ResolvedVisualizeBinding{})},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			err := test.handler(&cobra.Command{}, resolvedinput.Inputs{}, resolvedinput.Inputs{})
-			if err == nil || !strings.Contains(err.Error(), "service is required") {
-				t.Fatalf("handler error = %v, want required service error", err)
-			}
-		})
-	}
-}
-
-func TestResolvedApprovalAdaptersMapStableInputsIntoFreshRequests(t *testing.T) {
-	var listRequests []workcli.ListHumanApprovalsConfig
-	list := commandregistry.ResolvedApprovalListRunE(commandregistry.ResolvedApprovalListBinding{
-		ListHumanApprovals: func(cfg workcli.ListHumanApprovalsConfig) error {
-			listRequests = append(listRequests, cfg)
-			return nil
-		},
-		DiagnosticsWriter: func(cmd *cobra.Command) io.Writer { return cmd.ErrOrStderr() },
-	})
-	var showRequests []workcli.ShowHumanApprovalConfig
-	show := commandregistry.ResolvedApprovalShowRunE(commandregistry.ResolvedApprovalShowBinding{
-		ShowHumanApproval: func(cfg workcli.ShowHumanApprovalConfig) error {
-			showRequests = append(showRequests, cfg)
-			return nil
-		},
-		DiagnosticsWriter: func(cmd *cobra.Command) io.Writer { return cmd.ErrOrStderr() },
-	})
-
-	executeResolvedApprovalList(t, list, []string{
-		"--server", "https://factory.example", "--json", "--debug",
-		"work", "approval", "list", "--session", "session-alpha",
-	}, io.Discard, io.Discard, context.Background())
-	executeResolvedApprovalShow(t, show, []string{
-		"--server", "https://factory.example", "--json", "--debug",
-		"work", "approval", "show", "approval-alpha", "--session", "session-alpha",
-	}, io.Discard, io.Discard, context.Background())
-
-	if len(listRequests) != 1 || listRequests[0].Server != "https://factory.example" ||
-		listRequests[0].SessionID != "session-alpha" || !listRequests[0].JSON ||
-		!listRequests[0].Verbose || !listRequests[0].Debug || listRequests[0].Output == nil ||
-		listRequests[0].Diagnostics == nil || listRequests[0].Context == nil {
-		t.Fatalf("approval list config = %#v, want resolved request values", listRequests)
-	}
-	if len(showRequests) != 1 || showRequests[0].Server != "https://factory.example" ||
-		showRequests[0].SessionID != "session-alpha" || showRequests[0].ApprovalID != "approval-alpha" ||
-		!showRequests[0].JSON || !showRequests[0].Verbose || !showRequests[0].Debug ||
-		showRequests[0].Output == nil || showRequests[0].Diagnostics == nil || showRequests[0].Context == nil {
-		t.Fatalf("approval show config = %#v, want resolved request values", showRequests)
-	}
-}
-
-func TestResolvedApprovalAdaptersRejectMissingResolvedInputs(t *testing.T) {
-	list := commandregistry.ResolvedApprovalListRunE(commandregistry.ResolvedApprovalListBinding{
-		ListHumanApprovals: func(workcli.ListHumanApprovalsConfig) error { return nil },
-	})
-	if err := list(&cobra.Command{}, resolvedinput.Inputs{}, resolvedinput.Inputs{}); err == nil ||
-		!strings.Contains(err.Error(), "resolve human approval list inputs") {
-		t.Fatalf("approval list missing local inputs error = %v", err)
-	}
-	if err := list(&cobra.Command{}, resolvedTestInputs(t,
-		resolvedTestValue{id: "you.work.approval.list.flag.session", source: resolvedinput.SourceCLIFlag, value: resolvedinput.StringValue("session-alpha")},
-	), resolvedinput.Inputs{}); err == nil || !strings.Contains(err.Error(), "resolve human approval list inputs") {
-		t.Fatalf("approval list missing inherited inputs error = %v", err)
-	}
-
-	show := commandregistry.ResolvedApprovalShowRunE(commandregistry.ResolvedApprovalShowBinding{
-		ShowHumanApproval: func(workcli.ShowHumanApprovalConfig) error { return nil },
-	})
-	if err := show(&cobra.Command{}, resolvedinput.Inputs{}, resolvedinput.Inputs{}); err == nil ||
-		!strings.Contains(err.Error(), "resolve human approval show inputs") {
-		t.Fatalf("approval show missing local inputs error = %v", err)
-	}
-	if err := show(&cobra.Command{}, resolvedTestInputs(t,
-		resolvedTestValue{id: "you.work.approval.show.arg.0", source: resolvedinput.SourcePositionalArgument, value: resolvedinput.StringValue("approval-alpha")},
-		resolvedTestValue{id: "you.work.approval.show.flag.session", source: resolvedinput.SourceCLIFlag, value: resolvedinput.StringValue("session-alpha")},
-	), resolvedinput.Inputs{}); err == nil || !strings.Contains(err.Error(), "resolve human approval show inputs") {
-		t.Fatalf("approval show missing inherited inputs error = %v", err)
-	}
-}
-
-func executeResolvedApprovalList(
-	t *testing.T,
-	list commandregistry.ResolvedWorkRunE,
-	args []string,
-	output io.Writer,
-	diagnostics io.Writer,
-	ctx context.Context,
-) {
-	t.Helper()
-	if err := executeResolvedApprovalListError(t, list, args, output, diagnostics, ctx); err != nil {
-		t.Fatalf("Execute(%v) error = %v", args, err)
-	}
-}
-
-func executeResolvedApprovalListError(
-	t *testing.T,
-	list commandregistry.ResolvedWorkRunE,
-	args []string,
-	output io.Writer,
-	diagnostics io.Writer,
-	ctx context.Context,
-) error {
-	t.Helper()
-	noop := func(*cobra.Command, resolvedinput.Inputs, resolvedinput.Inputs) error { return nil }
-	root, err := climanifestcobra.NewResolvedWorkCommandTree(commandregistry.ResolvedWorkHandlers{
-		ApprovalList: list, ApprovalShow: noop, List: noop, Show: noop, Move: noop, Visualize: noop,
-	})
-	if err != nil {
-		t.Fatalf("NewResolvedWorkCommandTree() error = %v", err)
-	}
-	root.SetOut(output)
-	root.SetErr(diagnostics)
-	root.SetContext(ctx)
-	root.SetArgs(args)
-	return root.Execute()
-}
-
-func executeResolvedApprovalShow(
-	t *testing.T,
-	show commandregistry.ResolvedWorkRunE,
-	args []string,
-	output io.Writer,
-	diagnostics io.Writer,
-	ctx context.Context,
-) {
-	t.Helper()
-	if err := executeResolvedApprovalShowError(t, show, args, output, diagnostics, ctx); err != nil {
-		t.Fatalf("Execute(%v) error = %v", args, err)
-	}
-}
-
-func executeResolvedApprovalShowError(
-	t *testing.T,
-	show commandregistry.ResolvedWorkRunE,
-	args []string,
-	output io.Writer,
-	diagnostics io.Writer,
-	ctx context.Context,
-) error {
-	t.Helper()
-	noop := func(*cobra.Command, resolvedinput.Inputs, resolvedinput.Inputs) error { return nil }
-	root, err := climanifestcobra.NewResolvedWorkCommandTree(commandregistry.ResolvedWorkHandlers{
-		ApprovalList: noop, ApprovalShow: show, List: noop, Show: noop, Move: noop, Visualize: noop,
-	})
-	if err != nil {
-		t.Fatalf("NewResolvedWorkCommandTree() error = %v", err)
-	}
-	root.SetOut(output)
-	root.SetErr(diagnostics)
-	root.SetContext(ctx)
-	root.SetArgs(args)
-	return root.Execute()
-}
-
 type fixedHTTPClock struct{}
 
 func (fixedHTTPClock) Now() time.Time { return time.Unix(1, 0) }
