+++ pkg/transports/cli/root.go
@@ -109,6 +109,58 @@ type MoveWorkOperation func(workcli.MoveConfig) error
 type VisualizeWorkOperation func(workcli.VisualizeConfig) error
 type ListHumanApprovalsOperation func(workcli.ListHumanApprovalsConfig) error
 type ShowHumanApprovalOperation func(workcli.ShowHumanApprovalConfig) error
+
+func executeGeneratedHumanApprovalList(
+	cmd *cobra.Command,
+	globals *cliGlobalOptions,
+	diagnostics *cliDiagnosticsOptions,
+	list func(workcli.ListHumanApprovalsConfig) error,
+) error {
+	if list == nil {
+		return fmt.Errorf("human approval list service is required")
+	}
+	values, err := generatedCommandInputs(cmd)
+	if err != nil {
+		return err
+	}
+	sessionID, err := commandInputValue[string](values, "you.work.approval.list.flag.session")
+	if err != nil {
+		return err
+	}
+	return list(workcli.ListHumanApprovalsConfig{
+		Context: cmd.Context(), Server: globals.server, SessionID: sessionID,
+		JSON: globals.json, Output: cmd.OutOrStdout(),
+		Diagnostics: diagnostics.writer(cmd), Verbose: diagnostics.verboseEnabled(),
+		Debug: diagnostics.debug,
+	})
+}
+
+func executeGeneratedHumanApprovalShow(
+	cmd *cobra.Command,
+	args []string,
+	globals *cliGlobalOptions,
+	diagnostics *cliDiagnosticsOptions,
+	show func(workcli.ShowHumanApprovalConfig) error,
+) error {
+	if show == nil {
+		return fmt.Errorf("human approval show service is required")
+	}
+	values, err := generatedCommandInputs(cmd)
+	if err != nil {
+		return err
+	}
+	sessionID, err := commandInputValue[string](values, "you.work.approval.show.flag.session")
+	if err != nil {
+		return err
+	}
+	return show(workcli.ShowHumanApprovalConfig{
+		Context: cmd.Context(), Server: globals.server, SessionID: sessionID, ApprovalID: args[0],
+		JSON: globals.json, Output: cmd.OutOrStdout(),
+		Diagnostics: diagnostics.writer(cmd), Verbose: diagnostics.verboseEnabled(),
+		Debug: diagnostics.debug,
+	})
+}
+
 type ListWorkerSessionsOperation = workersessionscli.ListOperation
 type ShowWorkerSessionsOperation = workersessionscli.ShowOperation
 type ReadWorkerSessionOperation = workersessionscli.ReadOperation
