+++ pkg/transports/cli/work/show.go
@@ -15,3 +15,22 @@ func NewShow(transport clihttp.Protocol) func(ShowConfig) error {
 func Show(cfg ShowConfig) error {
 	return workcli.Show(cfg)
 }
+
+type ListHumanApprovalsConfig = workcli.ListHumanApprovalsConfig
+type ShowHumanApprovalConfig = workcli.ShowHumanApprovalConfig
+
+func NewListHumanApprovals(transport clihttp.Protocol) func(ListHumanApprovalsConfig) error {
+	return workcli.NewListHumanApprovals(transport)
+}
+
+func NewShowHumanApproval(transport clihttp.Protocol) func(ShowHumanApprovalConfig) error {
+	return workcli.NewShowHumanApproval(transport)
+}
+
+func ListHumanApprovals(cfg ListHumanApprovalsConfig) error {
+	return workcli.ListHumanApprovals(cfg)
+}
+
+func ShowHumanApproval(cfg ShowHumanApprovalConfig) error {
+	return workcli.ShowHumanApproval(cfg)
+}
