--- pkg/transports/http/test_helpers_test.go
@@ -1,5 +1,0 @@
-package http
-
-func httpStringPtr(value string) *string {
-	return &value
-}
